Stupid overflow in timer. Supposed to use a uint32_t, used a uint16_t by mistake. Caused no end of problems!
git-svn-id: svn+ssh://svn.code.sf.net/p/openlcb/svn/trunk/prototypes/C@1318 87536cb5-86f1-4cdc-8fc6-3f0f71bcd267

+++ libraries/OpenLCB/OLCB_Datagram_Handler.h
@@ -25,7 +25,6 @@ class OLCB_Datagram_Handler : public OLCB_Handler
   {
     if((*nid) == (*NID))
     {
-//      Serial.println("Someone wants to know my name!");
       return true;
     }
     return false;
@@ -45,7 +44,7 @@ class OLCB_Datagram_Handler : public OLCB_Handler
  protected:
  
  //TODO condense all these bools into a bitfield
-  uint16_t _sentTime;
+  uint32_t _sentTime;
   bool _initialized;
   bool _txFlag;
   bool _rxDatagramBufferFree;
